[darwin] Refactor single pid vs all pids for .ps
This is a refactoring of the `.ps` method to reduce the amount of extra
allocations necessary and unnecessary looping over unneeded pids when
`pid` is passed.

Effectively, this splits the method into pieces:

- If the method has a bid passed in, it will call get_info_for_pid, with
  the pid as the argument
- If the pid value is `nil`, then the get_info_for_all_pids method will
  be called, and get_info_for_pid will be called for each pid that it
  finds

This means that the proc_pidinfo method won't get called when a single
pid is requested for, saving a few object allocations.

All previous functionality should still be in place.