[darwin] Add .rss and .pss convenience methods
Generally, when `Sys::ProcTable` is used for a single process, usually
the user is only requesting a single value from all of the data
collected.  These methods are examples of using the recently refactored
`.ps` method's parts to access a smaller concise amount of data.  This
helps reduce the number of object allocations when the data needed is
minimal.