[darwin] Adds optional "lazy load" for exe parts
The exe, cmdline, and environ portions of the ProcTableStruct, from my
experience, tend to go unused by most use cases.  Instead of removing
that functionality by default, this addition of a optional `opts` param
of `:lazy` allows the user to get the exe, cmdline, and environ data on
request.

This has a bit of a dis-advantage as it will mean that a process could
die in the process of getting this data, so that might not be ideal.