Add bench_ips_ps script
Used this to test changes before and after to compare the number of
iterations per second a particular change had an effect on with `.ps`
being run.  Similar to the existing benchmarks/bench_ps.rb script, but
basically duplicates the runs, and labels 2 for the original, and two
for the source.

Sample invocation/usage:

    $ bench/bench_ips_ps --gem
    $ bench/bench_ips_ps --gem
    $ bench/bench_ips_ps --source
    $ bench/bench_ips_ps --source

Comparison results will be shown after the last run.  Do to limitations
in the benchmarks/ips hold functionality, this was the best way to run
this.