[linux] Reduce object allocations for smaps parse
This change takes advantage of a few things in ruby and how the smaps
file is structured to reduce the object allocations for a single
process* on linux.  These are:

- We can basically assume that the file will always look the same on all
  distros, and the existing regexp is accurate in describing that.
- When running ruby's `.to_i` on a string, it will ignore all
  whitespace, and then only process the next set of consecutive digits.
  So both `"   4000 kb".to_i` and `"4000.0123".to_i` would return 4000.
- `.start_with?` doesn't allocate any new strings in ruby

The current implementation of this will always have to generate
intermediate strings for the regexp matches, which seems to be 2 per
match (haven't counted though).

That said, while this method has less allocations, it doesn't seem to
perform noticeably faster, and it technically taking advantage of a ruby
quirk, which adds risk.

* gains of course are also applicable for `.ps(nil)` as well)